trng: Rename rand to trng_rand
The declaration for rand conflicts with the standard library declaration
so rename it from "rand" to "trng_rand". This has the benefit of making
it obvious when we're using the true random number generator.

For consistency, this also renames init_trng/exit_trng to
trng_init/trng_exit.

BRANCH=none
BUG=b:234181908
TEST=./util/compare_build.sh -b all -j 120
     => MATCH

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Ic3305a91263c45786c051eaa5b3689e7464aa0ab