trng: Rename rand to trng_rand
The declaration for rand conflicts with the standard library declaration
so rename it from "rand" to "trng_rand". This has the benefit of making
it obvious when we're using the true random number generator.

For consistency, this also renames init_trng/exit_trng to
trng_init/trng_exit.

This is a reland of commit a6b0b3554f59cc9b0c4aae9bff7dff075f2089a9.

BRANCH=none
BUG=b:234181908
TEST=./util/compare_build.sh -b all -j 120
     => MATCH
TEST=emerge-hatch ec-utils-test

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Ic26890572cb9865275c866b65b0532c5ab029865