ap_pwrseq: eSPI: Valid signal after first value received
Consider the virtual wire signal valid once the first value
has been received. It is observed that the bus reset
and virtual wire callbacks do sometimes interleave, so do not
reset the valid flag when a bus reset is received.

Also display whether a signal is invalid in the powerindebug output.

BUG=b:236561037
TEST=zmake build nivviks; flash & run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I261b59884ecce71678ca825d9d29d37195ac9691