ap_pwrseq: Refactor exit-hardoff handling
Redo the exit-hardoff handling so that the S5 inactivity timer
is extended to give the AP more time to handle a button press,
and also to clear the request flag when transitioning out of S5 to S4.

This is avoid the situation when the request-exit-hardoff flag
is set in S5, but the AP does not enter G3 because it is
restarting (e.g via a cool reboot).

BUG=b:234572032,b:235429065
TEST=Run firmware_ECPowerButton test
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I8039bf9a1c28ec24dfd313c52bbc0e2aeb7e6d59