ap_pwrseq: Clear exit-hard-off flag in S5
Clear the exit-hard-off flag when transitioning from
S5 to S4 in case it was requested when the AP was in S5.

This is checked and cleared in G3, but it is possible
for the power button handler to set this flag when the AP
is in S5, so it wasn't being cleared, and the next time
the AP went to G3, the AP would be immediately started.

BUG=b:234572032
TEST=Run firmware_ECPowerButton test
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I8039bf9a1c28ec24dfd313c52bbc0e2aeb7e6d59