zephyr: Ensure shell priority is higher than sysworkq
The default shell priority is calculated using
K_LOWEST_APPLICATION_THREAD_PRIO, which may eventuate to
be lower than the sysworkq, so override the shell thread priority
and ensure it is lower than the sysworkq priority.

Some console commands such as waitms rely on the shell not
being preempted by the sysworkq (which runs the watchdog feeder).

Add a build assert to verify.

BUG=b:236650959
TEST=zmake build nivviks; test firmware_ECWatchdog
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I29ef967b045414e6b5e71ca895b2528ad9400c32