test: drivers: fix test order dependency
The motionsense host command tests all depend of the [0]th sensor to
be considered active. This wasn't the case because the state of the
machine was previously defaulted to SENSOR_ACTIVE_S0_S3. When
shuffling it's possible that the EC will be in S5 during execution.
This change could either force the EC to be in S0/S3, or set the
active mask to include S5 (which is the approach this CL took)

BRANCH=none
BUG=b:233355058
TEST=zephyr.elf -seed=$(date +%s)

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: I5bc5e7ef1950b0be5217dc348c5b29c753db819d