test: Fix usb_alt_mode in-suite exec order flake
The tcpci_tcpc_alert function requires a short virtual wait within the
USB alternate mode test suite before connecting a port emulator in order
to ensure the partner emulator attaches to the port. This was made
apparent when adding a second test to the test suite.

Insert this wait in the before test suite handler of the usb alt mode test.

BRANCH=none
BUG=b:233100881
BUG=b:219562077
TEST=zmake test test-drivers

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: I8c7e78146e5f634644c36269cbb2d6eb6824274c