test: drivers: fix test order dependency
The motionsense host command depends on the sensor at index 0 to be
initialized. Normally this is done by the motion sense task, but we
can also do it via the console command. Add the call to the console
command accelinit to initialize the sensor before each test. Note:
the console command is tested in another suite so we can just assume
it works here.

BRANCH=none
BUG=b:233355058
TEST=zephyr.elf -seed=$(date +%s)

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: Icc28197be7a03c2ec269bd5961bc5d62fade320d