core/risc-rv32i: skip unsupport CSR on mt8188
The interrupt delegation CSR, medeleg and mideleg, are not implemented on
MT8188 RV55. Skip them for RV55 to prevent excepton.

The usage of these CSR in the boot code disables handling interrupt
outside of machine mode. This is the default behavior of RV55, so it's
safe to skip the access on RV55.

BRANCH=none
BUG=b:236331462
TEST=boot to scp console

Signed-off-by: Tinghan Shen <tinghan.shen@mediatek.com>
Change-Id: I11770ba2fb11147b3ee53736c583e7a964b9e7d7