ap_pwrseq: never mark virtual wires as invalid after startup
The eSPI specification states that virtual wires always have well-defined
states: on PLTRST some are reinitialized to a default value while others
are retained (which is handled by individual eSPI drivers), and on #RESET
pin assertion all virtual wires are reset to defined default values.

Power sequencing should only treat virtual wire inputs as invalid (with
indeterminate value) at startup, when it's possible the platform is already
running an in a non-default state and the EC needs to resynchronize VW
state with the platform.

The previous handling of channel readiness was incorrect because channel
readiness is a flow control mechanism, unrelated to VW values. Specifically
per eSPI specification 1.0 section 7.2.1.5, the Virtual Wire Channel Ready bit
indicates only whether the device is ready to accept transactions on the
channel. Even when the channel is disabled, "the state of all the Virtual
Wires must continue to be maintained internally."

BUG=b:235429065
TEST=With powerd stopped (to prevent graceful shutdown), a 14 second power
     button press on Nereid now goes to S5 and wakes to S0 on another press
     rather than getting stuck in S3 with SLP_S4 and SLP_S5 inputs marked
     invalid. AP power is still maintained correctly across sysjump.
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I20578e0631234fa8c7058f6a8f29b06cfa76b496