test/stdlib: Disable gcc stringop-truncation warning
gcc warns:

error: ‘__builtin_strncpy’ output truncated copying 10 bytes from a
string of length 12 [-Werror=stringop-truncation]

In this case the test is intentionally copying a subset of the string.

Also, since it's not obvious, add a note that when building this test
for the host we're actually testing the toolchain's C standard library,
not the EC "builtin" version.

BRANCH=none
BUG=b:234181908
TEST=./util/compare_build.sh -b all -j 120
     => MATCH

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I943de03ae62d7ae5e431809305ae68100e6da418