zephyr: test: split tests into separate binaries
This is a first pass at spliting up the tests into different binaries.
Doing this involved a small change in the zmake framework to pass the
project name into CMake. Then, the root CMakeLists is able to add
the correct subdirectory to the common sources. This will add that
set of tests as a zephyr library.

BRANCH=none
BUG=none
TEST=zmake test test-drivers-isl923x
TEST=zmake test test-drivers-led_driver

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: Ida49ab5c4f1ec156365c854251f548eac00d2c5e