ServoV4pX: Fix DP DiscSVID EnterMode NAK behavior
ServoV4p1 has capability to emulate a DP AltMode Adapter. This is
controlled via a CLI logical flag. An issue of contention is
behavior of a "NAK" when modal entry is not allowed / invalid.

As-implemented:
* NAK occurs at DiscID if DP is logically diabled
* NAK occurs at DiscMode if DTS pullups [CCD] are engaged
* (NAK occurs at DiscMode if unrecognized SVID is queried)

However, per USB-IF PD WG discussion, spec mandates:
* "DiscID" response is provided by AMA DUT regardless of state
* "DiscID" field "AltModes Supported" mandates "DiscMode" reply
* "DiscMode" is a fixed, invariant response regardless of state
* If a Mode is invalid due to state, the EnterMode is "NAK"ed

This CL corrects that oversight for ServoV4p1 and ServoV4p0.

Note: This is partial excerpt from abandoned major refactor 313b82b.

BUG=b:238142672
BRANCH=main
TEST=1. Connect PD analyzer to ServoVp41 in AMA emulation mode
     2. Configure DisplayPort to be [Disabled]
     2. Connect Chromebook
     3. Monitor PD trace for where NAK message occurs

Change-Id: If3044eb0dde98d93a49f61c7b08849fee5ae1a3d
Signed-off-by: Nathan K <nkolluru@google.com>