mt8186: reports forcing shutdown by long pwrkey press
When powerkey long pressed for 8(+-20%) seconds, the PMIC
will enter forcing shutdown mode and turn off the rail.
If PMIC enters the forcing shutdown process too fast,
then the shutdown reason report will be canceled, and we won't
know the exactly shutdown reason. This CL reports the shutdown reason
when the power is going for S0S3, and hasn't reported the reason.

BUG=none
TEST=1. at S0, ensure the long powerkey press reports 0x8009
     2. apshutdown works
     3. shutdown from AP works
     4. at S3, long press powerkey: S3->S0->S3->S5->G3
BRANCH=none

Change-Id: I1ea8e710a81ed7267c7f87c6c6d711a045603b7e
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>