bmi3xx: loop querying the written calibrated value
The critical path of writing the calibration value to
the delay for writing calibrated data will be added up to 240ms, and
this will exceed the EC/AP SPI communication timeout 200ms.

This CL loops the querying the result for an early return.

BUG=b:237953471
TEST=echo -n "0" \
> /sys/bus/iio/devices/iio\:device3/in_anglvel_x_calibbias
and doesn't complain the timeout.
BRANCH=none

Change-Id: Ic14e54c53d8bd5c0d83bd1f3eac20a17c98a1bdc
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3750539
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Huanhuan Liu <huanhuanl@chromium.org>