sm5803: disable GPADCs when in LPM
It's safe to disable the GPADCs when the SM5803 is not active and very
dangerous to disable them when it is active, but they consume about 7mW
of power that we want to save. Add support for disabling GPADCs, taking
care to never disable them when the charger is active.

This requires changes to users that inspect the VBUS voltage, because
analog VBUS readings (as used by the USB-C state machine to detect
when a source is connected) do not update when the VBUS GPADC is disabled
(as happens when the charger goes to runtime LPM). sm5803_get_vbus_voltage
is changed to return an error if the VBUS ADC is disabled and a new
function is added to use an alternate path if necessary, which is now
adopted by all users of the SM5803.

BUG=b:237697900
TEST=Idle chargers now disable GPADCs when in runtime LPM
     (i2c read_byte I2C_SUB_USB_C1_TCPC 0x31 1 => 0) on Nereid, and VBUS
     detection still works following CC detection (which did not work
     when ADCs were disabled without any further changes).
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I40ba9b80e5503b240cdf67025a2c33817a337adb