sm5803: disable GPADCs when in LPM
It's safe to disable the GPADCs when the SM5803 is not active and very
dangerous to disable them when it is active, but they consume about 7mW
of power that we want to save. Add support for disabling GPADCs, taking
care to never disable them when the charger is active.

BUG=b:237697900
TEST=failed: charger detection breaks when VBUS GPADC is disabled
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I40ba9b80e5503b240cdf67025a2c33817a337adb