ps8xxx: simplify Kconfig options for driver selection
Introduce a PLATFORM_EC_USB_PD_TCPM_PS8XXX option that enables the
ps8xxx driver, and have the options for individual chips select it. This
reduces the number of statements for each chip, and makes it clear
that the multi-chip support option is only relevant if at least one
chip is enabled.

BUG=b:236761058,b:237593618
TEST=Verified ps8xxx.c is still built on Nereid, and selected/implied
     options remain enabled with the PS8815 driver.
BRANCH=none

Change-Id: I6a2215e7496613fc5daada154497c2ba2ac5c44b
Signed-off-by: Peter Marheine <pmarheine@chromium.org>