test/printf: Fix invalid %p format test
When the size passed to crec_vsnprintf is 0, the function will
immediately return with -EC_ERROR_INVAL, which is not testing the code
path that these tests were intending. Change the size to the size of the
output buffer so the code in vfnprintf is executed. Note that
crec_vsnprintf always adds a terminating NUL, so the buffer is modified
even though the format was incorrect.

BRANCH=none
BUG=b:238433667, b:234181908
TEST=make buildall

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Iade22f224f7f946c3c345f47e0496cd3f1754b63