ec: Fix kconfig_check and run unit test in cq
Changed the unit test to fail if it can't find zephyr or the kconfiglib
instead of skipping, and make it pass.

Run the unit test in the CQ.

In the process, I discovered that it never used kconfiglib because of
errors with ZEPHYR_BASE not being set. Changed kconfig_check to output
the error when it falls back to no kconfiglib.

This exposed that there were missing configs in util/config_allowed.txt

BRANCH=None
BUG=b:238773780
TEST=make -j40 buildall && util/run_tests.sh

Change-Id: I28a050d448a40df034dd9f2305a2d17cd0797468
Signed-off-by: Jeremy Bettis <jbettis@google.com>