sensors: Ensure that correct sensor ID order is used
The motion sensor processing assumes that the sensor
order is accelerometers followed by gyroscopes, so force
the sensor ID order to match.

To identify a gyroscope sensor, a gyro property is added to
the gyroscope compatibles, and set to a default value.
The macros can then check for the existence of this property when
identifying a gyro sensor.

BUG=b:238139272
TEST=zmake build nivviks; check DTS
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I9157fe6f5e0541166614c5c5d54ad0960ab3a4dc