mkbp: don't queue mkbp events in S3
In S3, if a mkbp event is not a wake source, we should not queue it in
the mkbp fifo, otherwise the system will see a bunch of outdated event
after resume.

mkbp_fifo_add() uses the return value from mkbp_send_event() to
decide if it needs to queue the event. So we need to pass the decision
through the path activate_mkbp_with_events() -> mkbp_send_event() ->
mkbp_fifo_add().

BUG=b:238057993
TEST=suspend -> lidclose -> lidopen(=resume)
     verify that powerd does not see the lid close event.
BRANCH=cherry

Signed-off-by: Ting Shen <phoenixshen@google.com>
Change-Id: I3e1c58f97020d7ee2e3b4b56f14c4cadf51bef64