zephyr: tests: Fix bb_retimer test conflict with PD task
Some bb_retimer tests require to set a PD data role. When the PD task is
running, it is possible that the data role changes after the test sets
it. To prevent that bb_retimer tests that require specific PD data role
are run before tasks are started.

BUG=b:238880607
TEST=zmake -D test test-drivers
BRANCH=None

Signed-off-by: Tomasz Michalec <tm@semihalf.com>
Change-Id: Ia5f59a5c63f5089eeb4984672a786b96b0845877