run_device_tests.py: Fix Pylint error consider-using-with
Pylint recommends using with for the allocation of the
ThreadPoolExecutor object. Our script used the non-default shutsdown()
argument wait=False to try to avoid to hang at the end of the test
execution.

The additional argument is not actually preventing the hang from
happening, so it's okay to fix the Pylint error using "with" as
recommended.

BUG=b:239100048
TEST=cros lint test/run_device_tests.py
BRANCH=none

Signed-off-by: Andrea Grandi <agrandi@google.com>
Change-Id: Ia645221fa3630dfcc4fa9d9d1d1581fdb35797ca