dedede: limit PD voltage to 15V for boards with SM5803
Silicon Mitus have stated that the SM5803 may be damaged by voltage spikes
on VBUS in excess of 20V, and recommended that VBUS above 15V not be used
to prevent damage. Although dedede configures the max PD power to 45W which
will usually result in 15V@3A maximum, 20V@2.25A is also a valid choice
that would be dangerous.

Boards that do not use the SM5803 are unchanged, but the setting must
now be specified for each board because it is not uniform across all
dedede boards.

BUG=b:230712704
TEST=make buildall
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I5b2359fc2720ef0b5a7f2480ff67e0e821c378ad