util/getversion.sh: Fix empty file list handling
This fixes how an empty list of changed files is handled. Previously,
xargs would call 'ls -t' with no arguments when no files changed. This
is equivalent to 'ls -t .' which is not the intended behavior.  So,
simply tell xargs to not invoke 'ls -t' in the no-argument case and only
process non-zero length file names.

This would happen when git status returned a file status other than
'[M|A]' such as a delete.

BRANCH=none
BUG=none
TEST=make no longer complains when there are no changes

Change-Id: I01119ffa6777ec7d338f9426f816595c696cbe74
Signed-off-by: Caveh Jalali <caveh@chromium.org>