zephyr: Switch Twister tests to LLVM
Switch back to using the LLVM toolchain and use an existing wrapper
script for calling `llvm-cov gcov`. The wrapper script is necessary due
to `llvm-cov` requiring `gcov` as a subcommand and there is no way to
pass that in to Twister.

Also made two minor tweaks to `run_twister`:
 1. Turn extra_args in to an optional list param so multiple extra args
    can be specified
 2. Replace opt with a dedicated param for enabling code coverage

BRANCH=None
BUG=b:240093007
TEST=./twister; zephyr/firmware_builder.py

Signed-off-by: Tristan Honscheid <honscheid@google.com>
Change-Id: I9a4aa085eb71de1688d55b14b432b6c459d1a119