sm5803/nereid: trace runtime LPM on GPIO
This makes the GPL2 GPIO reflect the current power mode for the secondary
charger: it goes high when initialized and goes low when in LPM (as done in
response to a USB-PD disconnect), going high again on LPM exit.

GPIO timing is set such that it goes low when LPM entry begins, and goes
high when LPM exit completes.

BUG=b:230712704

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Iac06fa086f97ab29d598cfa3858ed96c1242b29a