it8xxx2: support USE_FLOATING_POINT
This change implements the EC-required floating-point functions in a way
that can be used on it8xxx2. It is an exception to the rule that CONFIG_FPU
is required for USE_FLOATING_POINT in zephyr, because the CPU does support
floating point but the compiler cannot be told due to tooling reasons.

BUG=b:227961840
TEST=???
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I751050772d516d6d1d9f04a905100d4f70edde0f