it8xxx2: permit usage of PLATFORM_EC_FPU
IT8xxx2 has an FPU but can't be supported by the compiler because
typical library configurations don't support F without M, and M
needs to be disabled for this chip due to a CPU erratum.

This change implements the EC floating-point functions in a way
that can be built for it8xxx2.

BUG=b:227961840
TEST=???
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I751050772d516d6d1d9f04a905100d4f70edde0f