nissa/nereid: use floating-point
IT8xxx2 has an FPU but we weren't using it. Enabling FPU support
should improve performance because if it's not available then the
motionsense code uses fixed-point which uses software emulation for
64-bit multiply and divide which is slow both because it requires
many instructions be executed and because the code is large so it
will tend to thrash the instruction cache.

BUG=b:227961840
TEST=Doesn't currently build
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I7afeacc3bf882b3b15cf47e7dd4c1eb0a2622e46