Decouple FPU presence from use of floating-point
Using CONFIG_FPU as a dual signal to both use floating-point types when
useful and to indicate that the CPU has a FPU makes it impossible to
enable use of software floating-point emulation, which is generally not
desirable due to poor performance and large code size but is useful
in some situations.

To allow use of floating-point emulation, add a CONFIG_SOFT_FLOAT
option to allow use of floating-point without CONFIG_FPU and better define
CONFIG_FPU to indicate that both a FPU is present and hardware
floating-point should be used (this matches the meaning of CONFIG_FPU in
Zephyr).

BUG=b:227961840
TEST=???
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Icdd5effb5faca116d8c22bf34674ac724002c74e