Decouple FPU presence from use of floating-point
Using CONFIG_FPU as a dual signal to both use floating-point types when
useful and to indicate that the CPU has a FPU makes it impossible to
enable use of software floating-point emulation, which is generally not
desirable due to poor performance and large code size but is useful
in some situations.

To allow use of floating-point emulation, add a CONFIG_USE_FLOATING_POINT
option to control whether floating-point is used or not and better define
CONFIG_FPU to indicate that both a FPU is present and hardware
floating-point should be used (this matches the meaning of CONFIG_FPU in
Zephyr).

In Zephyr, USE_FLOATING_POINT depends on FPU presence because no soft-float
implementation of fabsf() or sqrtf() is currently implemented.

TODO: can we default USE_FLOATING_POINT to FPU in legacy builds?

Change-Id: Icdd5effb5faca116d8c22bf34674ac724002c74e