zephyr: doc: add a note about alternate function GPIOs
On NPCX platforms some pins default to a specific (non GPIO) function
after reset, but are switched to GPIO in the initialization code. This
may be confusing if one is only referencing the manual default pin
multiplexing configuration.

Adding a note about that in the GPIO doc page to reference this
behavior.

BRANCH=none
BUG=b:241734677
TEST=check the page on gitiles

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I3f6fa9f93c339771e15ac5b62c53442b28d1d2ba