zephyr: tests: Allow passing EC and module dirs via env var
The Gitlab coverage reports are run without a chroot, so we cannot rely
on components needed for the build to be in their usual spot. This CL
modifies the wrapper script to allow specifying the EC directory and
Zephyr modules directory through environment variables that the Gitlab
CI script already exports.

Tested by leaving the chroot and passing in ZEPHYR_BASE, EC_DIR, and
MODULES_DIR via env vars.

BUG=b:240093007
BRANCH=None
TEST=./twister

Signed-off-by: Tristan Honscheid <honscheid@google.com>
Change-Id: I1cb225e414ed0dbc3f9a69516d90dc91f4c1127f