it8xxx2: implement EC sqrtf/absf
This change implements the EC-required floating-point functions in a way
that can be used on it8xxx2. It is unusual in that a general software
version of these functions is not implemented, but this one permits
soft-float be used on this chip without a general software implementation
of the functions.

BUG=b:227961840
TEST=Nereid now builds if CONFIG_PLATFORM_EC_SOFT_FLOAT=y
     (with code size increased ~4kB due to inclusion of assorted
     soft-float functions from libgcc).
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I751050772d516d6d1d9f04a905100d4f70edde0f