mt8186: Stay at S3S5 until PMIC off
Currently, PMIC is not fully off at S5 state, and we also treat the S5 as
a transition state, and after the PMIC is off, we then turn to G3.

However, this might mislead the users that S5 is a stable state.
Instead, we should wait for the PMIC off at S3S5 and until the PMIC off,
then we transit to S5 or G3.

BUG=b:242012415
TEST=1. stop daisydog; echo > /dev/watchdog
        wdt reset is triggered.
     2. in S0: dut-control power_key:8.2
        stay at S3S5 for 8 seconds, and then transit to S5
     3. in S0: apshutdown in EC console
        stay at S3S5 for 8 seconds, and then transit to S5
     4. in S0: poweroff in EC console
        transit to S5, G3 directly.
BRANCH=none

Change-Id: I49e1a4a30dbff29af426a61fcc3e25784e283ec0
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>