mt8186/mt8188: force shutdown AP by turning off PMIC
Previously, we manually asserts AP reset, and then hold the PMIC EN
long enough (8 secs) to bring down PMIC for a gracefully reset.
The 8 seconds hold of PMIC EN is the same as the requirement in
https://chromium.googlesource.com/chromiumos/docs/+/HEAD/debug_buttons.md.

This cause some issues that before entering S5, the PMIC
might sometimes still not on, and hence the SHUTDOWN COMPLETE HOOK
only be called until S5.

We now turn off the PMIC first, and then bring down the AP. This ensures
the PMIC is off first, so the SHUTDOWN COMPLETE HOOK can be called
int S3S5. It also ensures the EN_PP4200_S5 can be turned off later than
the PMIC off.

BUG=b:242012415 b:267268982
TEST=On Steelix, Tentacruel and Geralt:
     * Cold reset:
       $ dut-control cold_reset:on sleep:0.2 cold_reset:off
       Result: G3 -> S0
     * Long power press to shutdown:
       $ dut-control dut-control power_key:8.2
       Result: S0 -> S5 -> G3
     * Long power press to power-on but then shutdown:
       $ dut-control dut-control power_key:9.2
       Result: G3 -> S0 -> S5 -> G3
     * Short power press to power-on:
       $ dut-control dut-control power_key:tab
       Result: G3 -> S0
     * Console command: apreset
       Result: S0 -> S0, AP reboots
     * Console command: apshutdown
       Result: S0 -> S5 -> G3
     * Lid open to power-on:
       $ dut-control lid_open:no sleep:0.2 lid_open:yes
       Result: G3 -> S0
     * AP console: reboots
       Reulst: S0 -> S0
     * AP console: poweroff
       Reulst: S0 -> G3
     * Short power press to power-on:
       $ dut-control dut-control power_key:tab
       Result: G3 -> S0
BRANCH=none

Change-Id: Iacaa3dbcdafd61b2f3371e2ba376ebdcf29659ff
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>